Convert more tests to jest
